+++ adi-front-design-system/packages/docs/src/stories/Components/Form.stories.tsx
@@ -1,5 +1,5 @@
-import type { FormProps } from '@adi/react-components';
-import { Form } from '@adi/react-components';
+import type { FormProps } from '@it-adi/react-components';
+import { Form } from '@it-adi/react-components';
 import { Meta, StoryObj } from '@storybook/react';
 
 export default {