+++ adi-front-design-system/packages/docs/src/stories/Components/Input.stories.tsx
@@ -1,5 +1,5 @@
-import type { InputProps } from '@adi/react-components';
-import { Input } from '@adi/react-components';
+import type { InputProps } from '@it-adi/react-components';
+import { Input } from '@it-adi/react-components';
 import { Meta, StoryObj } from '@storybook/react';
 
 export default {