+++ adi-front-design-system/packages/docs/src/stories/Components/Title.stories.tsx
@@ -1,5 +1,5 @@
-import type { TitleProps } from '@adi/react-components';
-import { Title } from '@adi/react-components';
+import type { TitleProps } from '@it-adi/react-components';
+import { Title } from '@it-adi/react-components';
 import { Meta, StoryObj } from '@storybook/react';
 
 export default {