+++ adi-front-design-system/packages/docs/src/stories/Components/Tooltip.stories.tsx
@@ -1,5 +1,5 @@
-import type { TooltipProps } from '@adi/react-components';
-import { Tooltip } from '@adi/react-components';
+import type { TooltipProps } from '@it-adi/react-components';
+import { Tooltip } from '@it-adi/react-components';
 import { Meta, StoryObj } from '@storybook/react';
 
 export default {