+++ adi-front-design-system/packages/react-charts/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @adi/react-charts
+# @it-adi/react-charts
+
+## 1.0.2
+
+### Patch Changes
+
+- Alterado nome de dependências
 
 ## 1.0.1