+++ adi-front-design-system/packages/ts-config/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @adi/ts-config
+# @it-adi/ts-config
+
+## 1.0.2
+
+### Patch Changes
+
+- Alterado nome de dependências
 
 ## 1.0.1